+++ packages/twenty-server/project.json
@@ -208,14 +208,14 @@
       "executor": "nx:run-commands",
       "options": {
         "cwd": "packages/twenty-server",
-        "command": "nx ts-node-no-deps -- src/database/clickhouse/migrations/run-migrations.ts"
+        "command": "nx ts-node-no-deps -- src/database/clickHouse/migrations/run-migrations.ts"
       }
     },
     "clickhouse:seed": {
       "executor": "nx:run-commands",
       "options": {
         "cwd": "packages/twenty-server",
-        "command": "nx ts-node-no-deps -- src/database/clickhouse/seeds/run-seeds.ts"
+        "command": "nx ts-node-no-deps -- src/database/clickHouse/seeds/run-seeds.ts"
       }
     },
     "lingui:extract": {
