+++ packages/twenty-server/src/database/clickHouse/seeds/run-seeds.ts
@@ -2,7 +2,7 @@
 import { createClient } from '@clickhouse/client';
 import { config } from 'dotenv';
 
-import { fixtures } from 'src/engine/core-modules/audit/utils/fixtures/fixtures';
+import { fixtures } from './fixtures';
 
 config({
   path: process.env.NODE_ENV === 'test' ? '.env.test' : '.env',
